Change the group ID and rename the sub modules
