+++ gradle.properties
@@ -1,8 +1,8 @@
 org.gradle.caching=true
-group=org.ballerinalang
+group=io.ballerina.stdlib
 version=0.8.0-beta.3-SNAPSHOT
 
-ballerinaGradlePluginVersion=0.9.1
+ballerinaGradlePluginVersion=0.10.0
 
 ballerinaLangVersion=2.0.0-beta.2.1
-stdlibIoVersion=0.6.0-beta.2
+stdlibIoVersion=0.6.0-beta.3-20210708-142400-de75ca2
