--- os-test-utils/build.gradle
@@ -1,83 +1,0 @@
-/*
- * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-apply plugin: "java"
-apply plugin: "com.github.spotbugs"
-apply plugin: "checkstyle"
-
-description = 'Ballerina - OS Test Utils'
-
-dependencies {
-    implementation group: 'org.ballerinalang', name: 'ballerina-runtime', version: "${ballerinaLangVersion}"
-    implementation project(":os-native")
-    implementation group: 'org.apache.commons', name: 'commons-lang3', version: "${apacheCommonsVersion}"
-    checkstyle project(":build-config:checkstyle")
-    checkstyle "com.puppycrawl.tools:checkstyle:${puppycrawlCheckstyleVersion}"
-}
-
-spotbugsMain {
-    ignoreFailures = true
-    effort = "max"
-    reportLevel = "low"
-    reportsDir = file("$project.buildDir/reports/spotbugs")
-    def excludeFile = file("${rootDir}/build-config/spotbugs-exclude.xml")
-    if (excludeFile.exists()) {
-        it.excludeFilter = excludeFile
-    }
-    reports {
-        text.enabled = true
-    }
-}
-
-spotbugsTest {
-    enabled = false
-}
-
-task validateSpotbugs() {
-    doLast {
-        if (spotbugsMain.reports.size() > 0 &&
-                spotbugsMain.reports[0].destination.exists() &&
-                spotbugsMain.reports[0].destination.text.readLines().size() > 0) {
-            spotbugsMain.reports[0].destination?.eachLine {
-                println 'Failure: ' + it
-            }
-            throw new GradleException("Spotbugs rule violations were found.");
-        }
-    }
-}
-
-checkstyle {
-    toolVersion '7.8.2'
-    configFile file("${rootDir}/build-config/checkstyle/build/checkstyle.xml")
-    configProperties = ["suppressionFile": file("${rootDir}/build-config/checkstyle/build/suppressions.xml")]
-}
-
-tasks.withType(Checkstyle) {
-    exclude '**/module-info.java'
-}
-
-spotbugsMain.finalizedBy validateSpotbugs
-checkstyleMain.dependsOn(":build-config:checkstyle:downloadMultipleFiles")
-
-compileJava {
-    doFirst {
-        options.compilerArgs = [
-                '--module-path', classpath.asPath,
-        ]
-        classpath = files()
-    }
-}
